+++ src/main/kotlin/io/github/t45k/part/entity/MethodHistory.kt
@@ -5,8 +5,8 @@ import org.eclipse.jdt.core.dom.SingleVariableDeclaration
 
 data class MethodHistory(val fileName: String, val revisions: List<Revision>)
 
-data class Revision(val commitHash: String, val parameters: List<SingleVariableDeclaration>, val body: Block, val commitMessage: String)
+data class Revision(val parameters: List<SingleVariableDeclaration>, val body: Block, val commitMessage: String)
 
 data class RawMethodHistory(val fileName: String, val rawRevisions: List<RawRevision>)
 
-data class RawRevision(val commitHash: String, val commitMessage: String, val rawBody: String)
+data class RawRevision(val rawBody: String, val commitMessage: String)
