+++ src/main/kotlin/io/github/t45k/part/mining/git/GitCommand.kt
@@ -1,10 +1,10 @@
 package io.github.t45k.part.mining.git
 
+import org.eclipse.jgit.internal.storage.file.FileRepository
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
-import java.nio.file.Path
 
-abstract class GitCommand<Input, Output>(protected val projectRootPath: Path) {
+abstract class GitCommand<Input, Output>(protected val repository: FileRepository) {
     protected val logger: Logger = LoggerFactory.getLogger(this::class.java)
     abstract fun execute(input: Input): Output
 }
